+++ packages/nhsuk-frontend/src/nhsuk/common/index.mjs
@@ -130,9 +130,31 @@ export function isSupported($scope = document.body) {
 }
 
 /**
+ * Check for an object
+ *
+ * @template {Partial<Record<keyof ObjectType, unknown>>} ObjectType
+ * @param {unknown | ObjectType} option - Option to check
+ * @returns {option is ObjectType} Whether the option is an object
+ */
+export function isObject(option) {
+  return !!option && typeof option === 'object' && !Array.isArray(option)
+}
+
+/**
+ * Check for valid scope
+ *
+ * @template {Element | Document} ScopeType
+ * @param {unknown | ScopeType} $scope - Scope of the document to search within
+ * @returns {$scope is ScopeType} Whether the scope can be queried
+ */
+export function isScope($scope) {
+  return !!$scope && ($scope instanceof Element || $scope instanceof Document)
+}
+
+/**
  * Format error message
  *
- * @param {ComponentConstructor} Component - Component that threw the error
+ * @param {CompatibleClass} Component - Component that threw the error
  * @param {string} message - Error message
  * @returns {string} - Formatted error message
  */
@@ -143,5 +165,5 @@ export function formatErrorMessage(Component, message) {
 export * from './nhsuk-frontend-version.mjs'
 
 /**
- * @import { ComponentConstructor } from '../component.mjs'
+ * @import { CompatibleClass } from '../component.mjs'
  */