+++ packages/nhsuk-frontend/src/nhsuk/components/checkboxes/checkboxes.mjs
@@ -1,3 +1,4 @@
+import { normaliseOptions } from '../../common/configuration/index.mjs'
 import { toggleConditionalInput } from '../../common/index.mjs'
 import { Component } from '../../component.mjs'
 import { ElementError } from '../../errors/index.mjs'
@@ -11,7 +12,7 @@ import { ElementError } from '../../errors/index.mjs'
  */
 export class Checkboxes extends Component {
   /**
-   * @param {Element | null} [$root] - HTML element to use for component
+   * @param {Element | null} $root - HTML element to use for component
    */
   constructor($root) {
     super($root)
@@ -199,16 +200,21 @@ export class Checkboxes extends Component {
 /**
  * Initialise checkboxes component
  *
- * @param {object} [options]
- * @param {Element | Document | null} [options.scope] - Scope of the document to search within
+ * @deprecated Use {@link createAll | `createAll(Checkboxes)`} instead.
+ * @param {InitOptions} [options]
  */
-export function initCheckboxes(options = {}) {
-  const $scope = options.scope ?? document
-  const $checkboxes = $scope.querySelectorAll(
+export function initCheckboxes(options) {
+  const { scope: $scope } = normaliseOptions(options)
+
+  const $checkboxes = $scope?.querySelectorAll(
     `[data-module="${Checkboxes.moduleName}"]`
   )
 
-  $checkboxes.forEach(($root) => {
+  $checkboxes?.forEach(($root) => {
     new Checkboxes($root)
   })
 }
+
+/**
+ * @import { createAll, InitOptions } from '../../index.mjs'
+ */