+++ packages/nhsuk-frontend/src/nhsuk/components/tabs/tabs.jsdom.test.mjs
@@ -139,6 +139,7 @@ describe('Tabs', () => {
     })
 
     it('should throw with missing $root element', () => {
+      // @ts-expect-error Parameter '$root' not provided
       expect(() => new Tabs()).toThrow(
         `${Tabs.moduleName}: Root element (\`$root\`) not found`
       )